+++ src/.internal/core/Root.ts
@@ -18,7 +18,6 @@ import { Rectangle } from "./render/Rectangle";
 import { Tooltip } from "./render/Tooltip";
 import { NumberFormatter } from "./util/NumberFormatter";
 import { DateFormatter } from "./util/DateFormatter";
-import { DurationFormatter } from "./util/DurationFormatter";
 import { ILocale, Language } from "./util/Language";
 import { Events, EventDispatcher } from "./util/EventDispatcher";
 import { DefaultTheme } from "../themes/DefaultTheme";
@@ -163,11 +162,6 @@ export class Root implements IDisposer {
 	 */
 	public dateFormatter: DateFormatter = DateFormatter.new(this, {});
 
-	/**
-	 * @ignore
-	 * @todo will use at some point in the future
-	 */
-	public durationFormatter: DurationFormatter = DurationFormatter.new(this, {});
 
 	// Accessibility